requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.45.0
datasets>=2.12.0
evaluate>=0.4.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
accelerate>=0.26.0

